requirements.txt
flask
flask-cors
pandas
numpy
flask-talisman
python-dotenv
gunicorn
pytest
scikit-learn==1.3.0
